refactor(socks5): dedup V4/V6 DNS race-resolution helper
