+++ rama-socks5/src/client/proxy_connector.rs
@@ -25,7 +25,6 @@ use rama_net::{
 };
 use rama_utils::macros::{define_inner_service_accessors, generate_set_and_with};
 use std::net::IpAddr;
-use tokio::sync::mpsc;
 
 #[derive(Debug, Clone, Default)]
 /// A [`Layer`] which wraps the given service with a [`Socks5ProxyConnector`].
@@ -238,70 +237,7 @@ impl<S> Socks5ProxyConnector<S> {
                         }
                     }
                     DnsResolveIpMode::Dual | DnsResolveIpMode::DualPreferIpV4 => {
-                        use tracing::{Instrument, trace_span};
-
-                        let (tx, mut rx) = mpsc::unbounded_channel();
-
-                        tokio::spawn(
-                            {
-                                let tx = tx.clone();
-                                let domain = domain.clone();
-                                let dns_resolver = dns_resolver.clone();
-                                async move {
-                                    match dns_resolver.lookup_ipv4_rand(domain.clone()).await {
-                                        Some(Ok(addr)) => {
-                                            if let Err(err) = tx.send(IpAddr::V4(addr)) {
-                                                tracing::debug!(
-                                                    "failed to send ipv4 lookup result for ip: {addr}; err = {err:?}"
-                                                )
-                                            }
-                                        },
-                                        Some(Err(err)) => {
-                                            tracing::debug!(
-                                                "failed to lookup ipv4 addresses for domain: {err:?}"
-                                            );
-                                        }
-                                        None => {
-                                            tracing::debug!(
-                                                "failed to lookup ipv4 addresses for domain: no addresses found"
-                                            );
-                                        }
-                                    }
-                                }
-                            }
-                            .instrument(trace_span!("dns::ipv4_lookup")),
-                        );
-
-                        tokio::spawn(
-                            {
-                                let domain = domain.clone();
-                                let dns_resolver = dns_resolver.clone();
-                                async move {
-                                    match dns_resolver.lookup_ipv6_rand(domain.clone()).await {
-                                        Some(Ok(addr)) => {
-                                            if let Err(err) = tx.send(IpAddr::V6(addr)) {
-                                                tracing::debug!(
-                                                    "failed to send ipv6 lookup result for ip: {addr}; err = {err:?}"
-                                                )
-                                            }
-                                        },
-                                        Some(Err(err)) => {
-                                            tracing::debug!(
-                                                "failed to lookup ipv6 addresses for domain: {err:?}"
-                                            );
-                                        }
-                                        None => {
-                                            tracing::debug!(
-                                                "failed to lookup ipv6 addresses for domain: no addresses found"
-                                            );
-                                        }
-                                    }
-                                }
-                            }
-                            .instrument(trace_span!("dns::ipv6_lookup")),
-                        );
-
-                        rx.recv()
+                        crate::dns::race_resolve_dual(dns_resolver, domain.clone())
                             .await
                             .map(Host::Address)
                             .unwrap_or(Host::Name(domain))
